refactor: fold Conversation event handling into Conversation.swift
Keep incoming-event dispatch next to the session lifecycle so the
behavior is easier to find than in a separate extension file.
